% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Extract coefficients from linreg object}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments (not used)}
}
\value{
Named vector of coefficients
}
\description{
Extract coefficients from linreg object
}
