% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{Extract coefficients from ridgereg object}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{A ridgereg object}

\item{...}{Additional arguments (not used)}
}
\value{
Named vector of coefficients
}
\description{
Extract coefficients from ridgereg object
}
