% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Function}
\usage{
linreg(formula, data, method = "ols")
}
\arguments{
\item{formula}{A formula object specifying the model}

\item{data}{A data.frame containing the data}

\item{method}{Method to use: "ols" for ordinary least squares or "qr" for QR decomposition}
}
\value{
An object of class "linreg" containing regression results
}
\description{
Performs linear regression analysis using matrix computations
}
\examples{
data(iris)
mod <- linreg(Petal.Length ~ Species, data = iris)
print(mod)

}
