% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot method for linreg objects}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object}

\item{...}{Additional arguments (not used)}
}
\value{
A list containing two ggplot objects, the Residuals vs Fitted and Scale-Location plots
}
\description{
Creates two diagnostic plots using ggplot2, one plotting the residuals versus
fitted values and the scale-location plot (the square root of the
standardized residuals force the fitted values).
}
