% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{pred}
\alias{pred}
\title{Extract predicted values from linreg object}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{A linreg object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of fitted values
}
\description{
Extract predicted values from linreg object
}
