% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{predict}
\alias{predict}
\title{Extract predicted values from ridgereg object}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{A ridgereg object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of fitted values
}
\description{
Extract predicted values from ridgereg object
}
