% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Extract predicted values from ridgereg object}
\usage{
\method{predict}{ridgereg}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{A ridgereg object}

\item{...}{Additional arguments (not used)}
}
\value{
Vector of fitted values
}
\description{
Extract predicted values from ridgereg object
}
