% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg_methods.R
\name{print.linreg}
\alias{print.linreg}
\title{Print method for linreg objects
Prints the call the object was created with and the calculated regression coefficients.}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{A linreg object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for linreg objects
Prints the call the object was created with and the calculated regression coefficients.
}
