% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print method for ridgereg objects
Prints the call the object was created with and the calculated regression coefficients.}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{A ridgereg object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for ridgereg objects
Prints the call the object was created with and the calculated regression coefficients.
}
