% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge Regression Function}
\usage{
ridgereg(formula, data, lambda)
}
\arguments{
\item{formula}{A formula object specifying the model}

\item{data}{A data.frame containing the data}

\item{lambda}{A numeric scalar}
}
\value{
An object of class "ridgereg" containing regression results
}
\description{
Performs ridge regression analysis using least squares
}
\examples{
data(iris)
mod <- ridgereg(Petal.Length ~ Species, data = iris, lambda=0)
print(mod)

}
