% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg_methods.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Airport Delays}
\usage{
visualize_airport_delays()
}
\value{
A ggplot2 plot object
}
\description{
Creates a plot that visualizes the mean delay of flights for different airports
by longitude and latitude using ggplot2 and dplyr
}
